Lightsaber (She's Struggling)
Okay so I added a sound effect that makes it sound like a lightsaber humming.
And I added a feature when you are in edit mode the saber is red but when you click play it turns blue
Sadly I am still struggling with getting the saber to be seen as an item and therefore it won't be picked up by the player. I skimmed through the video again but couldn't figure out what I did wrong.
So I figured I would turn this in and hope that maybe I can update it later.