Gravity Gun Animation and Tower collapse
The gravity gun did not turn out great, but the functions are still there. Pressing e will start the idle animation. The primary i just change the barrels color to blue when spawning the cubes and secondary the barrels turn yellow when using the gravity on the cubes.  Collapse tower is just something i tried to replicate from the lecture. There is a new script i made for the Gravity gun animation, which might need some work on.